+++ rep_score_portal/views/asset_overview.py
@@ -1,6 +1,25 @@
+import warnings
+
 import pandas as pd
+import st_aggrid
 import streamlit as st
 
+from config import (
+    AGENCY_CREATIVE_LABEL_1,
+    AGENCY_CREATIVE_LABEL_2,
+    AGENCY_CREATIVE_LABEL_3,
+    AGENCY_CREATIVE_LABEL_4,
+    AGENCY_CREATIVE_LABEL_5,
+    DEI_CREATIVE_REVIEWS_LABEL_1,
+    DEI_CREATIVE_REVIEWS_LABEL_2,
+    DEI_CREATIVE_REVIEWS_LABEL_3,
+    DEI_CREATIVE_REVIEWS_LABEL_4,
+    DEI_CREATIVE_REVIEWS_LABEL_5,
+    MARKETING_LABEL_1,
+    MARKETING_LABEL_2,
+    MARKETING_LABEL_3,
+    MARKETING_LABEL_4,
+)
 from utils import (
     display_progress_bar_asset_tracker,
     edit_colors_of_selectbox,
@@ -15,6 +34,121 @@ def home_page() -> None:
 
     fetch_asset_data()
 
+    tab_1, tab_2 = st.tabs(['Asset Information', 'Progress of All Available Assets'])
+
+    with tab_1:
+        view_asset_information()
+    with tab_2:
+        view_progress_of_all_available_assets()
+
+    if not st.session_state.get('hacky_experimental_rerun_for_asset_overview_first_page_load'):
+        st.session_state.hacky_experimental_rerun_for_asset_overview_first_page_load = True
+        st.experimental_rerun()
+
+
+def view_asset_information() -> None:
+    """View all details and uploaded notes about a specific version of an assigned asset."""
+    asset_selected = st.selectbox(
+        label="Select an asset whose details you'd like to view",
+        options=['-'] + sorted(st.session_state.asset_tracker_df['Asset Name'].unique()),
+    )
+
+    if asset_selected and asset_selected != '-':
+        df_to_display = (
+            st.session_state.asset_tracker_df[
+                st.session_state.asset_tracker_df['Asset Name'] == asset_selected
+            ]
+            .rename(columns={'Username': 'Submitted By'})
+        )
+
+        if len(df_to_display) == 0:
+            st.error('Hmm... something went wrong here. Sorry, please refresh and try again.')
+            st.stop()
+
+        cols_to_display = [
+            'Asset Name',
+            'Date Submitted',
+            'Version',
+            'Submitted By',
+            'Status',
+            'Brand',
+            'Product',
+            'Region(s) This Creative Will Air In',
+            'Content Type',
+        ]
+
+        grid_options = st_aggrid.grid_options_builder.GridOptionsBuilder.from_dataframe(
+            dataframe=df_to_display[cols_to_display],
+        )
+        grid_options.configure_selection(selection_mode='single')
+        grid_options = grid_options.build()
+
+        # random, but... ¯\_(ツ)_/¯
+        height = 41 + (len(df_to_display) * 28)
+
+        st.caption('Click a row below to view more details about that asset.')
+
+        with warnings.catch_warnings():
+            # oof
+            warnings.filterwarnings('ignore', module=r'.*st_aggrid*')
+
+            data = st_aggrid.AgGrid(
+                data=df_to_display,
+                gridOptions=grid_options,
+                update_mode=st_aggrid.shared.GridUpdateMode.SELECTION_CHANGED,
+                height=height,
+            )
+
+        if len(data['selected_rows']) > 0:
+            row_selected = data['selected_rows'][0]
+
+            non_note_cols_to_display = [
+                'Submitted By',
+                'Version',
+                'Brand',
+                'Product',
+                'Region(s) This Creative Will Air In',
+                'Content Type',
+                'Point of Contact Email',
+                'Date Submitted',
+            ]
+
+            note_cols_to_display = [
+                MARKETING_LABEL_1,
+                MARKETING_LABEL_2,
+                MARKETING_LABEL_3,
+                MARKETING_LABEL_4,
+                AGENCY_CREATIVE_LABEL_1,
+                AGENCY_CREATIVE_LABEL_2,
+                AGENCY_CREATIVE_LABEL_3,
+                AGENCY_CREATIVE_LABEL_4,
+                AGENCY_CREATIVE_LABEL_5,
+                DEI_CREATIVE_REVIEWS_LABEL_1,
+                DEI_CREATIVE_REVIEWS_LABEL_2,
+                DEI_CREATIVE_REVIEWS_LABEL_3,
+                DEI_CREATIVE_REVIEWS_LABEL_4,
+                DEI_CREATIVE_REVIEWS_LABEL_5,
+                'Notes',
+            ]
+
+            st.markdown(body=f'### {row_selected["Asset Name"]}')
+
+            for col in non_note_cols_to_display:
+                st.markdown(body=f'**{col}**: {row_selected[col]}')
+
+            insert_line_break()
+
+            for col in note_cols_to_display:
+                st.markdown(body=f'**{col if col != "Notes" else "Notes:"}**')
+
+                if row_selected[col] and row_selected[col] != 'N/A':
+                    st.text(body=f'{row_selected[col]}')
+                else:
+                    st.text(body='N/A')
+
+
+def view_progress_of_all_available_assets() -> None:
+    """View coding progress of all pending and assigned uploaded assets."""
     if st.session_state.asset_information.get('name'):
         st.markdown('### In Progress Assets')